Data Query Enabler v2.4
DEMETER Pilot 5.2 - Use Case 1. This component enables data extraction through a query interface directly connected to the spatio-temporal semantic data management system of the AFarCloud smart farming platform. In turn, the component performs a runtime translation of the data for representation and integration into the Knowage platform. This service provides information on the observations generated by static and mobile sensors and devices in a determined area. Thanks to the query interface, the management of large volumes of spatio-temporal data are enabled.